Fix Express Checkout is not active by default when plugin is installed
The main issue is, the hook `upgrader_process_complete` where the
`migrateSharedOptions` and `activateExpressCheckout` actions are
hooked runs correctly but that because the old plugin already contains
those actions, when the new plugin is installed the code of that
new version will run only after the installation means adding an hook
will not be executed because WordPress do the action `upgrader_process_complete`
in an HTTP request where our hooks are added to WordPress after the redirection so in another HTTP request making impossible for the plugin to run the code.

I can fix it with a temporary hook but this is something we have to solve
for all other plugins, the first time we want to run an upgrade process
for a plugin that doesn't contains already functions hooked to it we have
to run the upgrade process to every plugin bootstrap.

Why this is a problem, because for every page load we have to do
additional queries to the database and not having something already
to execute code during an upgrade means we don't know when we'll could
remove the secondary approach. Can't be in the version 2.0.1 nor in the
version 2.1.0 because we don't know which users and when they'll
do the upgrade.

To summarize, for 2.0.0 I'll include both, hooks in `upgrader_process_complete`
so in future versions we can just use them and in meantime I'll execute
the same functions after the plugin has been bootstrapped.

This solve also the issue with the shared options.

PPP-279

+++ src/Install/Installer.php
@@ -46,11 +46,20 @@ class Installer
     }
 
     /**
+     * Perform Tasks After Plugin is Installed or Upgraded
+     */
+    public function afterInstall()
+    {
+        $this->migrateSharedOptions();
+        $this->activateExpressCheckout();
+    }
+
+    /**
      * Migrate Shared options
      *
      * @return void
      */
-    public function migrateSharedOptions()
+    private function migrateSharedOptions()
     {
         $options = get_option(self::ORIGINAL_OPTIONS, []);
 
@@ -64,7 +73,7 @@ class Installer
     /**
      * Activate Express Checkout Gateway When Plugin get Installed
      */
-    public function activateExpressCheckout()
+    private function activateExpressCheckout()
     {
         $enabled = $this->expressCheckoutGateway->get_option('enabled', 'no');
         $enabled = wc_string_to_bool($enabled);
@@ -73,6 +82,40 @@ class Installer
             return;
         }
 
+        if (!method_exists($this->expressCheckoutGateway, 'update_option')) {
+            $this->gatewayUpdateOption('enabled', 'yes');
+            return;
+        }
+
         $this->expressCheckoutGateway->update_option('enabled', 'yes');
+    }
+
+    /**
+     * Backward Compatibility Method Because `WC_Settings_API::update_option` does not Exists
+     * Prior to WooCommerce 3.4.x
+     *
+     * @param $key
+     * @param null $empty_value
+     * @return mixed
+     */
+    private function gatewayUpdateOption($key, $empty_value = null)
+    {
+        if (empty($this->expressCheckoutGateway->settings)) {
+            $this->expressCheckoutGateway->init_settings();
+        }
+
+        // Get option default if unset.
+        if (!isset($this->expressCheckoutGateway->settings[$key])) {
+            $form_fields = $this->expressCheckoutGateway->get_form_fields();
+            $this->expressCheckoutGateway->settings[$key] = isset($form_fields[$key])
+                ? $this->expressCheckoutGateway->get_field_default($form_fields[$key])
+                : '';
+        }
+
+        if (!is_null($empty_value) && $this->expressCheckoutGateway->settings[$key] === '') {
+            $this->expressCheckoutGateway->settings[$key] = $empty_value;
+        }
+
+        return $this->expressCheckoutGateway->settings[$key];
     }
 }
